Chore: cleaning up Home greeting
